requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
faiss-cpu==1.12.0
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kagglehub==0.3.13
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruff==0.14.4
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
